% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMOptimFunction.R
\name{SVM.optim}
\alias{SVM.optim}
\title{Implement the SVM function using optim()}
\usage{
SVM.optim(x, y, cost, gradient, C, method)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix of the classes}

\item{cost}{the cost function}

\item{gradient}{the gradient function}

\item{C}{the hyperparameter C for SVM}

\item{method}{the method to use in the optim() function. Here, we suggest L-BFGS-B or Nelder-Mead}
}
\description{
Implement the SVM function using optim()
}
